AI Job Search Agent - API/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-dotenv
orjson

langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma

tavily-python
chromadb
sentence-transformers

pypdf
python-docx

numpy